shadcn-ui/requirements.txt
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2
langchain>=0.0.200
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
wandb>=0.15.0
python-dotenv>=1.0.0
physionet-client>=2.0.0
openai>=0.27.0
faiss-cpu>=1.7.4
pinecone-client>=2.2.2
pytorch-lightning>=2.0.0
accelerate>=0.20.0
datasets>=2.12.0
